update: move services/fdb_conf under services/fdb
