+++ cmd/sanssh/main.go
@@ -48,7 +48,6 @@ import (
 	_ "github.com/Snowflake-Labs/sansshell/services/sansshell/client"
 	_ "github.com/Snowflake-Labs/sansshell/services/service/client"
 	//_ "github.com/Snowflake-Labs/sansshell/services/fdb/client" // To get FDB modules uncomment this line.
-	//_ "github.com/Snowflake-Labs/sansshell/services/fdb_conf/client" // To get FDB config modules uncomment this line.
 )
 
 const (
