--- services/fdb_conf/fdb_conf.proto
@@ -1,55 +1,0 @@
-/* Copyright (c) 2019 Snowflake Inc. All rights reserved.
-
-   Licensed under the Apache License, Version 2.0 (the
-   "License"); you may not use this file except in compliance
-   with the License.  You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing,
-   software distributed under the License is distributed on an
-   "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-   KIND, either express or implied.  See the License for the
-   specific language governing permissions and limitations
-   under the License.
-*/
-
-syntax = "proto3";
-
-option go_package = "github.com/Snowflake-Labs/sansshell/services/fdb_conf";
-
-import "google/protobuf/empty.proto";
-
-package FdbConf;
-
-// The FdbConf service definition.
-service FdbConf {
-    // Read a configuration value from a section in FDB config file.
-    rpc Read(ReadRequest) returns (FdbConfResponse) {}
-    // Write updates a configuration value in a section of FDB config file.
-    rpc Write(WriteRequest) returns (google.protobuf.Empty) {}
-    // Delete a section value based on a key
-    rpc Delete(DeleteRequest) returns (google.protobuf.Empty) {}
-}
-
-message Location {
-    string file = 1;
-    string section = 2;
-    string key = 3;
-}
-
-message ReadRequest {
-    Location location = 1;
-}
-
-message WriteRequest {
-    Location location = 1;
-    string value = 2;
-}
-
-message DeleteRequest {
-    Location location = 1;
-}
-
-//
-message FdbConfResponse { string value = 1; }
